requirements.txt
langchain
python-dotenv
langchain-openai
icecream
thefuzz
pandas
fastapi
uvicorn
langchain-community
streamlit